requirements.txt
Flask
openai
requests
flask_cors
python-dotenv
langdetect
